Renaming onnx.py to onnx_reader.py to prevent global import errors
